pyproject.toml
[project]
name = "vlmrag"
version = "0.1.0"
description = "Vision-Language Model RAG API"
requires-python = ">=3.12,<3.13"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
	{ name = "vlmrag" }
]
dependencies = [
	"fastapi==0.115.4",
	"uvicorn[standard]==0.32.0",
	"pydantic==2.9.2",
	"python-dotenv==1.0.1",
	"requests==2.32.3",
	"numpy==2.1.3",
	"Pillow==11.0.0",
	"qdrant-client==1.12.1",
	"openai==1.54.3",
	"torch==2.5.1",
	"open_clip_torch==2.26.1",
	"PyMuPDF==1.24.11",
	"python-multipart==0.0.17",
]

[project.optional-dependencies]
dev = [
	"ruff==0.6.9",
	"pytest==8.3.3",
]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-select = ["I", "F", "E", "W"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
